Introduce EnterObserver.
